Improve course view. Split up code
