+++ WidgetExtension/Extensions/Bundle+appBundle.swift
@@ -1,0 +1,24 @@
+//
+//  Created for xikolo-ios under MIT license.
+//  Copyright © HPI. All rights reserved.
+//
+
+import Foundation
+
+extension Bundle {
+
+    static let appBundle: Bundle = {
+        var bundle = Bundle.main
+
+        if bundle.bundleURL.pathExtension == "appex" {
+            // Peel off two directory levels - MY_APP.app/PlugIns/MY_APP_EXTENSION.appex
+            let url = bundle.bundleURL.deletingLastPathComponent().deletingLastPathComponent()
+            if let otherBundle = Bundle(url: url) {
+                bundle = otherBundle
+            }
+        }
+
+        return bundle
+    }()
+
+}
